Hide internal APIs from the barrel
