+++ packages/hierarchy-builder/src/hierarchy-builder.ts
@@ -2,7 +2,6 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export * from "./hierarchy-builder/Common";
 export * from "./hierarchy-builder/IQueryExecutor";
 export * from "./hierarchy-builder/TreeNode";
 export * from "./hierarchy-builder/TreeNodesProvider";
