+++ packages/hierarchy-builder/src/hierarchy-builder/ECSqlBinding.ts
@@ -3,50 +3,11 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import { QueryBinder } from "@itwin/core-common";
-
-/** @internal */
+/** @beta */
 export type ECSqlBindingType = "boolean" | "double" | "id" | "idset" | "int" | "long" | "string" | "point2d" | "point3d";
 
-/** @internal */
+/** @beta */
 export interface ECSqlBinding {
   type: ECSqlBindingType;
   value?: any;
-}
-
-/** @internal */
-export function bind(bindings: ECSqlBinding[]): QueryBinder {
-  const binder = new QueryBinder();
-  bindings.forEach((b, i) => {
-    switch (b.type) {
-      case "boolean":
-        binder.bindBoolean(i + 1, b.value);
-        break;
-      case "double":
-        binder.bindDouble(i + 1, b.value);
-        break;
-      case "id":
-        binder.bindId(i + 1, b.value);
-        break;
-      case "idset":
-        binder.bindIdSet(i + 1, b.value);
-        break;
-      case "int":
-        binder.bindInt(i + 1, b.value);
-        break;
-      case "long":
-        binder.bindLong(i + 1, b.value);
-        break;
-      case "point2d":
-        binder.bindPoint2d(i + 1, b.value);
-        break;
-      case "point3d":
-        binder.bindPoint3d(i + 1, b.value);
-        break;
-      case "string":
-        binder.bindString(i + 1, b.value);
-        break;
-    }
-  });
-  return binder;
 }
